+++ src/gmes/application/facade.py
@@ -6,6 +6,7 @@ from .migration_uc import migrate_credentials
 from .runtime_uc import (execute_data_forms, execute_data_read, execute_data_stream,
                          execute_guided_workflow, execute_login, execute_run,
                          execute_run_request)
+from .supervisor_uc import run_supervised
 from ..profiles.store import known as known_profiles, load as load_profile
 
 
@@ -17,4 +18,4 @@ def package_version():
 __all__ = ["execute_data_forms", "execute_data_read", "execute_data_stream", "execute_doctor",
            "execute_guided_workflow", "execute_login", "execute_run", "execute_run_request",
            "migrate_credentials", "known_profiles", "load_profile", "package_version",
-           "set_credentials"]
+           "run_supervised", "set_credentials"]